% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{approximation_type_mapper}
\alias{approximation_type_mapper}
\title{Plot the frequency of the true parameter being in the credible interval}
\usage{
approximation_type_mapper(approximation_type)
}
\arguments{
\item{results}{A list of results from the simulation containing the credible intervals}

\item{prior_types}{A character vector of prior types}

\item{approximation_types}{A character vector of approximation types}

\item{parameter_names}{A character vector of parameter names}

\item{path}{A character string of the path to save the plot}

\item{dpi}{An integer of the dpi to save the plot}

\item{width}{An integer of the width of the plot}

\item{height}{An integer of the height of the plot}

\item{titles}{A character vector of titles for each plot}

\item{txt_size}{An integer of the text size by default 20}
}
\description{
Plot the frequency of the true parameter being in the credible interval for each prior type and approximation type
}
