% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{bootstrap_pairwise_exchangeability}
\alias{bootstrap_pairwise_exchangeability}
\title{Bootstrap exchangeability}
\usage{
bootstrap_pairwise_exchangeability(data1, data2, B = 10000)
}
\arguments{
\item{B}{The number of bootstrap samples by default 10000
return The bootstrap estimate of the p-value. The probability that model 1 is better than model 2}

\item{score1}{The first score}

\item{score2}{The second score}
}
\description{
Given two prediction scores {s_1,...,s_n}, {t_1,...,t_n}, calculates bootstrap estimate of the p-value for pairwise exchangeability
}
