% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{bootstrap_q}
\alias{bootstrap_q}
\title{Bootstrap}
\usage{
bootstrap_q(data, B = 10000)
}
\arguments{
\item{data}{The data}

\item{B}{The number of bootstrap samples by default 10000
return The bootstrap estimate of the 95\% confidence interval of the mean}
}
\description{
Given data s_1,...,s_n, calculates bootstrap estimate of the 95\% confidence interval
}
