% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{DSS}
\alias{DSS}
\title{DSS}
\usage{
DSS(mu, Q, y)
}
\arguments{
\item{mu}{The posterior mean of the variable of interest}

\item{Q}{The posterior precision of the variable of interest}

\item{y}{The observed data}
}
\value{
The Dawid-Sebastiani score (smaller is better)
}
\description{
Calculates the Dawid-Sebastiani score
}
