% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{EG_prior_r}
\alias{EG_prior_r}
\title{Exponential gaussian prior on absolute value of v}
\usage{
EG_prior_r(alpha_v, a0)
}
\arguments{
\item{alpha_v}{A quantile in (0,1) for the absolute value of v}

\item{a0}{A surprisingly high ratio of anisotropy, in (1,infinity)}
}
\value{
A function for the the prior on the absolute value of v
}
\description{
Calculates  the prior on the absolute value of |v|, supposing vi~N(0,sigma^2)
}
