% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{inla_KS_copy}
\alias{inla_KS_copy}
\title{Plot KS statistics and p-values}
\usage{
inla_KS_copy(x, y, ...)
}
\arguments{
\item{results}{A list of results from the simulation containing the probabilities}

\item{prior_types}{A character vector of prior types}

\item{approximation_types}{A character vector of approximation types}

\item{parameter_names}{A character vector of parameter names}

\item{path1}{A character string of the path to save the plot for the KS statistics}

\item{path2}{A character string of the path to save the plot for the p-values}

\item{dpi}{An integer of the dpi to save the plot}

\item{width}{An integer of the width of the plot}

\item{height}{An integer of the height of the plot}
}
\description{
Plot KS statistics and p-values for each prior type and approximation type
}
