% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{interval_score}
\alias{interval_score}
\title{Interval score}
\usage{
interval_score(parameter, credible_intervals, alpha = 0.05)
}
\arguments{
\item{parameter}{A vector of length d representing the parameter}

\item{credible_intervals}{A matrix of size d x 2 representing the credible intervals}

\item{alpha}{The significance level. By default, set to 0.05.}
}
\value{
A vector of length d representing the interval score for each dimension
}
\description{
Calculates the interval score for a given parameter and credible intervals
}
