% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{KS_table}
\alias{KS_table}
\title{Tables for KS test}
\usage{
KS_table(ks_results)
}
\arguments{
\item{ks_results}{A data frame containing the results of the KS test when simulated from a single prior}
}
\description{
Create tables for KS test
}
