% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{log_pc_prior_iso}
\alias{log_pc_prior_iso}
\title{Log-prior density for isotropic parameters}
\usage{
log_pc_prior_iso(sigma_u0, sigma_epsilon0, rho0, alpha = 0.01)
}
\arguments{
\item{sigma_u0}{Quantile for standard deviation of the prior of the field u}

\item{sigma_epsilon0}{Quantile for standard deviation of the noise}

\item{rho0}{Quantile for the correlation range}

\item{alpha}{Quantile by default 0.01}
}
\description{
Calculates the log-prior density of parameters (log(kappa), log(sigma_u), log(sigma_epsilon)) given quantiles
}
