% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{log_post_u_precipitation}
\alias{log_post_u_precipitation}
\title{Log-posterior density of field u given a linear noisy observation y= A*u + beta_0 +beta_1 h + epsilon}
\usage{
log_post_u_precipitation(
  log_prior_theta,
  log_sigma_beta,
  mesh,
  y,
  A,
  h,
  m_theta = rep(0, 5)
)
}
\arguments{
\item{log_prior_theta}{A function that calculates the log prior of theta = (log(kappa), v, log(sigma_u), log(sigma_epsilon))}

\item{log_sigma_beta}{The log variance of the linear effect beta}

\item{mesh}{The mesh}

\item{y}{A vector with length equal to the number of columns of A representing the observed data.}

\item{A}{Matrix of size mxn where n is the number of basis elements representing the transformation A}

\item{h}{A numeric vector representing the covariate}

\item{m_theta}{Arbitrary point in calculation of the log-posterior. By default, set to 0}
}
\value{
The calculated log-posterior
}
\description{
Calculates the log-posterior density of field u given a linear noisy observation y= A*u + beta_0 +beta_1 h + epsilon
Only stationary parameters are accepted.
Value is up to an additive constant depending only on y
beta =(beta_0,beta_1) is the linear effect and is modelled to be Gaussian independent from the rest of the parameters
}
