% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{log_posterior_prior_precipitation}
\alias{log_posterior_prior_precipitation}
\title{Log-posterior density for precipitation model}
\usage{
log_posterior_prior_precipitation(
  log_prior_theta,
  log_sigma_beta,
  mesh,
  log_kappa,
  v = NULL,
  log_sigma_u = 0,
  log_sigma_epsilon,
  y,
  A,
  h,
  m_u = 0,
  m_beta = 0
)
}
\arguments{
\item{log_prior_theta}{A function that calculates the log prior of theta = (log(kappa), v, log(sigma_u), log(sigma_epsilon))}

\item{log_sigma_beta}{The log variance of the linear effect beta}

\item{mesh}{The mesh}

\item{log_kappa}{Logarithm of inverse correlation range}

\item{v}{2D vector that controls anisotropy}

\item{log_sigma_u}{Variance of field u. If unspecified, it is assumed to be 0.}

\item{log_sigma_epsilon}{Variance of noise}

\item{y}{A vector with length equal to the number of columns of A representing the observed data.}

\item{A}{Matrix of size mxn where n is the number of basis elements representing the transformation A}

\item{h}{A numeric vector representing the covariate}

\item{m_u}{A vector with length n representing the prior mean m_u. If a number is given, it is transformed into (m_u, m_u,..., m_u). By default, set to 0}

\item{m_beta}{A vector with length 2 representing the prior mean m_beta. If a number is given, it is transformed into (m_beta, m_beta). By default, set to 0}
}
\value{
The calculated log-posterior

The calculated log-posterior
}
\description{
Calculates  the log-posterior density of parameters (log(kappa),v, log(sigma_u), log(epsilon)))
given a linear noisy observation y= A*u + beta_0 +beta_1 h + epsilon
Only stationary parameters are accepted.
Value is up to an additive constant depending only on y
beta =(beta_0,beta_1) is the linear effect and is modelled to be Gaussian independent from the rest of the parameters
}
