% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{log_smoothed_importance_weights_and_samples}
\alias{log_smoothed_importance_weights_and_samples}
\title{Smoothed log importance weights}
\usage{
log_smoothed_importance_weights_and_samples(
  log_posterior_density,
  mu_Gaussian_median,
  Q_Gaussian_median,
  n_weights,
  alpha = 0.05
)
}
\arguments{
\item{log_posterior_density}{A function that calculates the log posterior density of the parameters}

\item{mu_Gaussian_median}{The median of the Gaussian importance density of the parameters}

\item{Q_Gaussian_median}{The precision matrix of the Gaussian importance density of the parameters}

\item{n_weights}{The number of importance weights}

\item{alpha}{The quantile for the credible interval, by default 0.05}
}
\value{
The smoothed log importance weights and samples
}
\description{
Calculates the smoothed self normalized weights where the importance density is a Gaussian distribution
}
