% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{LOH_scores}
\alias{LOH_scores}
\title{Leave out half testing}
\usage{
LOH_scores(
  log_prior_theta,
  log_sigma_beta,
  mesh,
  location_indexes,
  data,
  h_mesh,
  max_iterations_MAP,
  theta0,
  n_weights,
  path_data = NULL
)
}
\arguments{
\item{log_prior_theta}{A function that calculates the log prior of theta = (log(kappa), v, log(sigma_u), log(sigma_epsilon))}

\item{log_sigma_beta}{The log variance of the linear effect beta}

\item{mesh}{The mesh}

\item{location_indexes}{A vector of indices representing the training data. The testing data is the complement of the training data}

\item{data}{A list with x and y representing the observed data.}

\item{h_mesh}{A numeric vector representing the covariate at the mesh points}

\item{max_iterations_MAP}{Maximum number of iterations for MAP, by default 300}

\item{theta0}{Initial value for the parameters (log(kappa), v, log(sigma_u), log(sigma_epsilon)). By default, set to (-0.5, c(0.1, 0.1), 0, -3, 0, 0)}

\item{n_weights}{The number of importance weights, by default 1000}

\item{path_data}{Path to save the result}

\item{path_prediction_txt}{Path to save the prediction}
}
\value{
The calculated scores and estimates
}
\description{
Calculates the validation for the precipitation model using half the points for training and half for testing
}
