% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{LOH_scores_loop}
\alias{LOH_scores_loop}
\title{Leave out half loop}
\usage{
LOH_scores_loop(
  log_prior_theta,
  log_sigma_beta,
  mesh,
  data,
  location_indexes_matrix,
  max_iterations_MAP = 300,
  theta0 = c(-0.5, c(0.1, 0.1), 0, -3, 0, 0),
  n_weights = 1000,
  h_mesh,
  path_data = NULL,
  path_prediction_txt = NULL
)
}
\arguments{
\item{log_prior_theta}{A function that calculates the log prior of theta = (log(kappa), v, log(sigma_u), log(sigma_epsilon))}

\item{log_sigma_beta}{The log variance of the linear effect beta}

\item{mesh}{The mesh}

\item{data}{A list with x and y representing the observed data}

\item{location_indexes_matrix}{A matrix. Each column are the indices of the data that is observed}

\item{max_iterations_MAP}{Maximum number of iterations for MAP, by default 300}

\item{theta0}{Initial value for the parameters (log(kappa), v, log(sigma_u), log(sigma_epsilon)). By default, set to (-0.5, c(0.1, 0.1), 0, -3, 0, 0)}

\item{n_weights}{The number of importance weights, by default 1000}

\item{path_data}{Path to the data}

\item{path_prediction_txt}{Path to the prediction}
}
\value{
The calculated average of the leave out half scores
}
\description{
Calculates the validation for the precipitation model using half the points for training and half for testing
}
