% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{LOM_scores}
\alias{LOM_scores}
\title{Leave out multiple testing}
\usage{
LOM_scores(
  log_prior_theta,
  log_sigma_beta,
  mesh,
  y1,
  A1,
  h1,
  y2,
  A2,
  h2,
  max_iterations_MAP = 300,
  theta0 = c(-0.5, c(0.1, 0.1), 0, -3, 0, 0),
  n_weights = 1000,
  h_mesh,
  path_data = NULL,
  path_prediction_txt = NULL,
  alpha = 0.05
)
}
\arguments{
\item{log_prior_theta}{A function that calculates the log prior of theta = (log(kappa), v, log(sigma_u), log(sigma_epsilon))}

\item{log_sigma_beta}{The log variance of the linear effect beta}

\item{mesh}{The mesh}

\item{y1}{A vector with length equal to the number of columns of A1 representing the observed data.}

\item{A1}{Matrix of size m1xn1 where n1 is the number of basis elements representing the transformation A1}

\item{h1}{A numeric vector representing the covariate at the observed data}

\item{y2}{A vector with length equal to the number of columns of A2 representing the training data.}

\item{A2}{Matrix of size m2xn2 where n2 is the number of basis elements representing the transformation A2}

\item{h2}{A numeric vector representing the covariate at the training data}

\item{max_iterations_MAP}{Maximum number of iterations for MAP, by default 300}

\item{theta0}{Initial value for the parameters (log(kappa), v, log(sigma_u), log(sigma_epsilon)). By default, set to (-0.5, c(0.1, 0.1), 0, -3, 0, 0)}

\item{n_weights}{The number of importance weights, by default 1000}

\item{h_mesh}{A numeric vector representing the covariate at the mesh points}

\item{path_data}{Path to the data}

\item{path_prediction_txt}{Path to the prediction}

\item{alpha}{The quantile for the credible interval, by default 0.05}
}
\value{
The calculated leave out one cross validation
}
\description{
Calculates the validation for the precipitation model leaving out multiple points
}
