% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{LOO_CV_precipitation}
\alias{LOO_CV_precipitation}
\title{LOO for the precipitation model}
\usage{
LOO_CV_precipitation(
  log_prior_theta,
  log_sigma_beta,
  mesh,
  y,
  A,
  h,
  max_iterations_MAP = 300,
  theta0 = c(-0.5, c(0.1, 0.1), 0, -3, 0, 0),
  n_weights = 1000,
  h_mesh,
  path_data = NULL,
  path_prediction_txt = NULL,
  quiet = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{log_prior_theta}{A function that calculates the log prior of theta = (log(kappa), v, log(sigma_u), log(sigma_epsilon))}

\item{log_sigma_beta}{The log variance of the linear effect beta}

\item{mesh}{The mesh}

\item{A}{Matrix of size mxn where n is the number of basis elements representing the transformation A}

\item{h}{A numeric vector representing the covariate at the observed data}

\item{max_iterations_MAP}{Maximum number of iterations for MAP, by default 300}

\item{theta0}{Initial value for the parameters (log(kappa), v, log(sigma_u), log(sigma_epsilon)). By default, set to (-0.5, c(0.1, 0.1), 0, -3, 0, 0)}

\item{n_weights}{The number of importance weights, by default 1000}

\item{h_mesh}{A numeric vector representing the covariate at the mesh points}

\item{path_data}{Path to the data}

\item{path_prediction_txt}{Path to the prediction
param alpha The quantile for the credible interval, by default 0.05}

\item{Y}{A vector with length equal to the number of columns of A representing the observed data.}
}
\value{
The calculated leave out one cross validation
}
\description{
Calculates the leave out one cross validation for the precipitation model
}
