% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{parameters_to_non_log}
\alias{parameters_to_non_log}
\title{Parameters to non log parameters}
\usage{
parameters_to_non_log(theta)
}
\arguments{
\item{theta}{A vector of parameters}
}
\value{
A vector of non log parameters
}
\description{
Converts parameters to non log parameters
}
\examples{
theta <- c(-1, 0.1, 0.1, -3, 1)
parameters_to_non_log(theta)
}
