% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{PC_prior_kappa}
\alias{PC_prior_kappa}
\title{Marginal PC prior density of kappa}
\usage{
PC_prior_kappa(lambda, lambda1)
}
\arguments{
\item{lambda}{A hyperparameter controlling the penalization of the distance from the base model as a function of kappa}

\item{lambda1}{A hyperparameter controlling the penalization of the distance from the base model as a function of v}
}
\value{
The marginal prior density function of kappa .
}
\description{
Calculates  the marginal PC prior pi_kappa of the inverse correlation range kappa
}
