% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{pc_prior_kappa_iso}
\alias{pc_prior_kappa_iso}
\title{Prior density function for kappa if field is isotropic}
\usage{
pc_prior_kappa_iso(rho0, alpha = 0.01)
}
\arguments{
\item{rho0}{Quantile for the correlation range}

\item{alpha}{Quantile by default 0.01}
}
\description{
Returns the prior density function of parameters rho given quantiles
}
