% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{PC_prior_kappa_quantile}
\alias{PC_prior_kappa_quantile}
\title{Marginal PC prior density function of kappa}
\usage{
PC_prior_kappa_quantile(rho0, a0, alpha = 0.01)
}
\arguments{
\item{rho0}{Surprisingly small correlation range}

\item{a0}{Surprisingly large amount of anisotropy}

\item{alpha}{Quantile by default 0.01}
}
\value{
The marginal prior density function of kappa .
}
\description{
Calculates  the marginal PC prior density pi_kappa
}
