% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{PC_prior_r}
\alias{PC_prior_r}
\title{Marginal PC prior density of r = |v|}
\usage{
PC_prior_r(r, lambda1)
}
\arguments{
\item{lambda1}{A hyperparameter controlling the penalization of the distance from the base model as a function of r.}

\item{log_r}{The logarithm of |v|, which controls the magnitude of the anisotropy}
}
\value{
The calculated marginal prior density of log(r).
}
\description{
Calculates  the marginal PC prior, pi_r, of the log norm of the anisotropy vector
}
\examples{
r <- -0.5
lambda1 <- 1
result <- PC_prior_r(log_r, lambda1)
}
