% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{PC_prior_r_quantile}
\alias{PC_prior_r_quantile}
\title{Log PC prior on r using quantiles}
\usage{
PC_prior_r_quantile(a0, alpha)
}
\arguments{
\item{a0}{A surprisingly high ratio of anisotropy, in (1,infinity)}

\item{alpha}{A quantile in (0,1) for anisotropy}
}
\value{
The PC prior function on r
}
\description{
Calculates  the log of the PC prior on r given quantiles
}
