% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{pc_prior_rho}
\alias{pc_prior_rho}
\title{Prior density function for rho, anisotropic field}
\usage{
pc_prior_rho(rho0, a0, alpha = 0.01)
}
\arguments{
\item{rho0}{Surprisingly small correlation range}

\item{a0}{Surprisingly large amount of anisotropy}

\item{alpha}{Quantile by default 0.01}
}
\description{
Returns the prior density function of parameters rho given quantiles
}
