% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt_distances_to_MAP}
\alias{plt_distances_to_MAP}
\title{Plot distances to MAP}
\usage{
plt_distances_to_MAP(
  results_list,
  prior_types,
  path = NULL,
  dpi = 100,
  width = 10,
  height = 5,
  titles,
  labeller = latex_labeller,
  variable_indexes = c(1, 2, 3, 4, 5),
  txt_size = 20,
  lower_bound = 10^-2,
  txt_angle = 45
)
}
\arguments{
\item{results_list}{A list of results from the simulation containing the distances to MAP}

\item{prior_types}{A character vector of prior types}

\item{path}{A character string of the path to save the plot}

\item{dpi}{An integer of the dpi to save the plot. By default, it is 600}

\item{width}{An integer of the width of the plot. By default, it is 10}

\item{height}{An integer of the height of the plot. By default, it is 10}

\item{titles}{A character vector of titles for each plot}

\item{variable_indexes}{An integer vector of the indexes of the variables to plot. By default, it is c(1, 2, 3, 4, 5)}

\item{txt_size}{An integer of the text size. By default, it is 20}
}
\value{
A ggplot object
}
\description{
Plot distances to MAP for each prior type
}
