% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt_KL_and_get_mean_KL}
\alias{plt_KL_and_get_mean_KL}
\title{Plot KL divergences}
\usage{
plt_KL_and_get_mean_KL(
  results_list,
  prior_types,
  approximation_types,
  path = NULL,
  dpi = 100,
  width = 10,
  height = 10,
  titles,
  txt_size = 20
)
}
\arguments{
\item{prior_types}{A character vector of prior types}

\item{approximation_types}{A character vector of approximation types}

\item{path}{A character string of the path to save the plot}

\item{dpi}{An integer of the dpi to save the plot}

\item{width}{An integer of the width of the plot}

\item{height}{An integer of the height of the plot}

\item{titles}{A character vector of titles for each plot}

\item{txt_size}{An integer of the text size by default 20}

\item{results}{A list of results from the simulation containing the KL divergences}
}
\description{
Plot KL divergences for each prior type and approximation type
}
