% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt_marginal_variance_of_u}
\alias{plt_marginal_variance_of_u}
\title{Plot the marginal variance K(x,x) of the field}
\usage{
plt_marginal_variance_of_u(
  mesh,
  boundary,
  kappa_list,
  v_list,
  sigma_u_list,
  path = NULL,
  titles,
  n_col = 4,
  txt_size = 30
)
}
\arguments{
\item{mesh}{A mesh object}

\item{boundary}{Boundary of the mesh}

\item{kappa_list}{A list of values for kappa}

\item{v_list}{A list of vectors for v}

\item{sigma_u_list}{A list of values for the standard deviation sigma_u}

\item{path}{A character string of the path to save the plot}

\item{titles}{A character vector of titles for each plot}

\item{n_col}{An integer of the number of columns for the facet wrap}
}
\description{
Plot the covariance K(x,x) of the field with a point for a list of functions for kappa and v
}
