% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt_pc_prior_kappa_v}
\alias{plt_pc_prior_kappa_v}
\title{Plot the pc priors on kappa and v}
\usage{
plt_pc_prior_kappa_v(
  lambda,
  lambda1,
  alpha = 0.01,
  l = 4,
  n_points = 100,
  path = NULL,
  txt_size = 20
)
}
\arguments{
\item{lambda}{Controls the size of kappa}

\item{lambda1}{Controls the size of v}

\item{alpha}{The quantile}

\item{l}{The length of the partition}

\item{n_points}{The number of points in the partition}

\item{path}{A character string of the path to save the plot}

\item{txt_size}{An integer of the text size by default 20}
}
\description{
Plot the pc priors on kappa and v
}
