% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt_pc_priors_kappa_rho_aniso_iso}
\alias{plt_pc_priors_kappa_rho_aniso_iso}
\title{Plot the pc prior for r and rho}
\usage{
plt_pc_priors_kappa_rho_aniso_iso(
  rho0,
  a0,
  alpha = 0.01,
  l = 20,
  n_points = 1000,
  path = NULL,
  txt_size = 20
)
}
\arguments{
\item{rho0}{An unexpectedly small correlation range}

\item{a0}{An unexpectedly small anisotropy}

\item{alpha}{The quantile}

\item{l}{The length of the partition}

\item{n_points}{The number of points in the partition}

\item{path}{A character string of the path to save the plot}

\item{txtsize}{An integer of the text size by default 20}
}
\description{
Plot the pc prior for r and rho
}
