% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt_u_sample}
\alias{plt_u_sample}
\title{title Plot realizations of a field on a mesh}
\usage{
plt_u_sample(
  mesh,
  boundary,
  kappa_list,
  v_list,
  sigma_u_list,
  path = NULL,
  titles,
  n_col = 4,
  txt_size = 30
)
}
\arguments{
\item{mesh}{A mesh object}

\item{boundary}{Boundary of the mesh}

\item{kappa_list}{A list of functions for kappa}

\item{v_list}{A list of functions for v}

\item{sigma_u_list}{A list of values for the standard deviation sigma_u}

\item{path}{A character string of the path to save the plot}

\item{titles}{A character vector of titles for each plot}
}
\description{
Plot realizations of a list of fields on a mesh given a list of functions for kappa and v
}
