% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPrior.R
\name{vector_to_half_angle_vector}
\alias{vector_to_half_angle_vector}
\title{Half angle vector}
\usage{
vector_to_half_angle_vector(v)
}
\arguments{
\item{v}{A 2D vector}
}
\value{
The 2D half angle vector
}
\description{
Converts a vector to its half angle vector
}
